requirements.txt
requests
requests_ntlm
google-api-python-client
pytz
